requirements.txt
bs4==0.0.2
numpy==1.26.4
pandas==2.1.4
pendulum==3.0.0
pytz==2024.2
requests==2.32.3
selenium==4.27
SQLAlchemy==1.4.54
apache-airflow-providers-databricks
boto3